Refactor menus
