--- apps/web/app/[locale]/[event]/components/Menu/LifeEventsMenu.module.scss
@@ -1,9 +1,0 @@
-@import "/app/styles/globals.scss";
-
-.events {
-  @include phone {
-    width: 100vh;
-  }
-
-  width: 200px;
-}
