+++ apps/web/app/[locale]/layout.module.scss
@@ -7,7 +7,7 @@
   padding: 0 15px;
   box-sizing: border-box;
 
-  @include large-desktop {
+  @include wide-desktop {
     max-width: 1400px;
   }
 }
