+++ apps/web/app/[locale]/layout.tsx
@@ -12,8 +12,8 @@ import { getAllEnabledFlags, isFeatureFlagEnabled } from "feature-flags/utils";
 import {
   getEnabledOptions,
   menuOptions,
-} from "./[event]/components/Menu/options";
-import SidebarWrapper from "./[event]/components/SidebarWrapper";
+} from "../components/HamburgerMenu/options";
+import HamburgerMenuWrapper from "../components/HamburgerMenu/HamburgerMenuWrapper";
 import { getTranslations } from "next-intl/server";
 import styles from "./layout.module.scss";
 
@@ -25,6 +25,8 @@ export default async function RootLayout({
   params: { locale: string };
 }) {
   const { userId, firstName, lastName } = await PgSessions.get();
+
+  const userName = [firstName, lastName].join(" ");
 
   const result = await pgpool.query<{ isInitialized: boolean }>(
     `SELECT EXISTS (SELECT 1 FROM user_consents WHERE user_id = $1 AND agreement = 'storeUserData' LIMIT 1) AS "isInitialized"`,
@@ -40,28 +42,13 @@ export default async function RootLayout({
     redirect(url.href, RedirectType.replace);
   }
 
-  const showEventsMenu = await isFeatureFlagEnabled("eventsMenu");
+  const showHamburgerMenu = await isFeatureFlagEnabled("eventsMenu");
 
-  let SidebarComponent;
-  if (showEventsMenu) {
-    const userName = [firstName, lastName].join(" ");
-
-    const enabledEntries = await getAllEnabledFlags(
-      menuOptions.map((o) => o.key),
-    );
-    const t = await getTranslations("EventsMenu");
-
-    const options = getEnabledOptions(locale, enabledEntries, t);
-    SidebarComponent = (
-      <SidebarWrapper
-        userName={userName}
-        ppsn="TUV1234123"
-        selected={path || ""}
-        options={options}
-        locale={locale}
-      />
-    );
-  }
+  const enabledEntries = await getAllEnabledFlags(
+    menuOptions.map((o) => o.key),
+  );
+  const eventsMenuT = await getTranslations("EventsMenu");
+  const options = getEnabledOptions(locale, enabledEntries, eventsMenuT);
 
   return (
     <html lang={locale}>
@@ -74,8 +61,15 @@ export default async function RootLayout({
           flexDirection: "column",
         }}
       >
-        {SidebarComponent}
-        <Header showSidebarToggle={showEventsMenu} locale={locale} />
+        {showHamburgerMenu && (
+          <HamburgerMenuWrapper
+            userName={userName}
+            selected={path || ""}
+            options={options}
+            locale={locale}
+          />
+        )}
+        <Header showHamburgerButton={showHamburgerMenu} locale={locale} />
         {/* All designs are made for 1440 px  */}
         <main className={styles.main}>
           <FeedbackBanner />
