+++ apps/web/app/components/HamburgerMenu/HamburgerMenu.module.scss
@@ -1,0 +1,32 @@
+@import "/app/styles/globals.scss";
+
+.events {
+  @include phone {
+    width: 100vh;
+  }
+
+  width: 200px;
+}
+
+.backdrop {
+  position: absolute;
+  height: 100vh;
+  width: 100vw;
+  z-index: 200;
+  background-color: gray;
+  opacity: 0.3;
+  min-height: 1px;
+}
+
+.sidebar {
+  position: absolute;
+  height: 100%;
+  z-index: 300;
+  background-color: white;
+  transform: translateX(-100vh);
+  transition: all 0.3s;
+}
+
+.visible {
+  transform: translateX(0px);
+}
