+++ apps/web/app/components/Header.module.scss
@@ -61,7 +61,7 @@
   justify-content: center;
   align-items: center;
 
-  @media ((max-width: $breakpoint-tablet)) {
+  @media ((max-width: $breakpoint-phone)) {
     display: none;
   }
 }
